requirements.txt
# Django E-Commerce Requirements
# ShopPy - E-Commerce Order Processing System

# Core Framework
Django>=4.2,<5.0

# Database Adapters
mysqlclient>=2.2.0        # MySQL adapter for PythonAnywhere
psycopg2-binary>=2.9.9    # PostgreSQL adapter (if using PostgreSQL)
dj-database-url>=2.1.0    # Parse DATABASE_URL environment variable

# Static Files
whitenoise>=6.6.0         # Serve static files in production

# Image Processing
Pillow>=10.0.0

# Environment Variables
python-dotenv>=1.0.0

# Security
bcrypt>=4.0.0
